requirements.txt
asgiref==3.4.1
Django==4.0
django-email-users==1.1.0
django-widget-tweaks==1.4.9
Pillow==8.4.0
signals==0.0.2
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
